WIP: Animate screenshots when they're being shown
This is proof of concept to animate the screenshots. Rather than showing
the screenshot at the moment in time it will now animate through all the
screenshots.

This clearly isn't ideal but I just wanted to poke around a little. Some
other interface should be used to do this if it is desirable.